Correct casting from size_t to int.
git-svn-id: file:///home/jj/hercules.svn/trunk@6301 956126f8-22a0-4046-8f4a-272fa8102e63

+++ vmfplc2.c
@@ -582,7 +582,7 @@ struct TAPE_BLOCKS *load_text_file(char *infile,char recfm,int *recl,int *recc,s
     recs=initrecs(recfm,*recl,plcd_hdr,5);
     while((rec=fgets(bfr,sizeof(bfr),ifile))!=NULL)
     {
-        rsz=strlen(rec)-1;
+        rsz=(int)strlen(rec)-1;
         if(recfm=='F')
         {
             if(rsz<(*recl))
